initial setup
promisify twitch extension helper
